+++ pyjls/binding.pyx
@@ -556,6 +556,18 @@ cdef class Reader:
         rc = c_jls.jls_rd_utc(self._rd, signal_id, sample_id, _utc_cbk_fn, <void *> cbk_fn)
         _handle_rc('rd_utc', rc)
 
+    def sample_id_to_timestamp(self, signal_id, sample_id):
+        cdef int64_t utc_timestamp
+        rc = c_jls.jls_rd_sample_id_to_timestamp(self._rd, signal_id, sample_id, &utc_timestamp)
+        _handle_rc('sample_id_to_timestamp', rc)
+        return utc_timestamp
+
+    def timestamp_to_sample_id(self, signal_id, utc_timestamp):
+        cdef int64_t sample_id
+        rc = c_jls.jls_rd_timestamp_to_sample_id(self._rd, signal_id, utc_timestamp, &sample_id)
+        _handle_rc('timestamp_to_sample_id', rc)
+        return sample_id
+
 
 cdef int32_t _annotation_cbk_fn(void * user_data, const c_jls.jls_annotation_s * annotation):
     obj: AnnotationCallback = <object> user_data