+++ src/rd_fsr.c
@@ -21,6 +21,7 @@
 #include <stddef.h>
 #include <stdlib.h>
 #include <inttypes.h>
+#include <math.h>
 
 
 #define ENTRIES_ALLOC_INIT  (1000)
@@ -60,7 +61,7 @@ struct jls_rd_fsr_s * jls_rd_fsr_alloc(double sample_rate) {
     return s;
 }
 
-void jls_rd_fsr_utc_free(struct jls_rd_fsr_s * self) {
+void jls_rd_fsr_free(struct jls_rd_fsr_s * self) {
     if (NULL != self) {
         if (NULL != self->sample_id) {
             free(self->sample_id);
@@ -76,7 +77,7 @@ void jls_rd_fsr_utc_free(struct jls_rd_fsr_s * self) {
     }
 }
 
-int32_t jls_rd_fsr_add(struct jls_rd_fsr_s * self, int64_t sample_id, int64_t utc) {
+int32_t jls_rd_fsr_add(struct jls_rd_fsr_s * self, int64_t sample_id, int64_t timestamp) {
     int64_t * p1;
     int64_t * p2;
     if (self->entries_length >= self->entries_alloc) {
@@ -102,8 +103,16 @@ int32_t jls_rd_fsr_add(struct jls_rd_fsr_s * self, int64_t sample_id, int64_t ut
         }
     }
     self->sample_id[self->entries_length] = sample_id;
-    self->utc[self->entries_length] = utc;
+    self->utc[self->entries_length] = timestamp;
     ++self->entries_length;
+    return 0;
+}
+
+int32_t jls_rd_fsr_add_cbk(void * user_data, const struct jls_utc_summary_entry_s * utc, uint32_t size) {
+    struct jls_rd_fsr_s * self = (struct jls_rd_fsr_s *) user_data;
+    for (uint32_t i = 0; i < size; ++i) {
+        jls_rd_fsr_add(self, utc[i].sample_id, utc[i].timestamp);
+    }
     return 0;
 }
 
@@ -132,35 +141,39 @@ int64_t interp_i64(struct jls_rd_fsr_s * self, int64_t x0, int64_t const * x, in
     double ds = (double) (x[low + 1] - x[low]);
     double dt = (double) (y[low + 1] - y[low]);
     double slope = dt / ds;
-    int64_t k = (int64_t) (dk * slope);
+    int64_t k = (int64_t) round(dk * slope);
     return y[low] + k;
 }
 
-int64_t jls_rd_fsr_sample_id_to_utc(struct jls_rd_fsr_s * self, int64_t sample_id) {
+int32_t jls_rd_fsr_sample_id_to_timestamp(struct jls_rd_fsr_s * self, int64_t sample_id, int64_t * timestamp) {
     if (self->entries_length == 0) {
-        return 0;
+        return JLS_ERROR_UNAVAILABLE;
     } else if (self->entries_length == 1) {
         if (self->sample_rate <= 0) {
-            return 0;
+            return JLS_ERROR_UNAVAILABLE;
         }
         double dsample = (double) (sample_id - self->sample_id[0]);
         double dt = dsample / self->sample_rate;
         dt *= JLS_TIME_SECOND;
-        return self->utc[0] + (int64_t) dt;
+        *timestamp = self->utc[0] + (int64_t) dt;
+    } else {
+        *timestamp = interp_i64(self, sample_id, self->sample_id, self->utc);
     }
-    return interp_i64(self, sample_id, self->sample_id, self->utc);
+    return 0;
 }
 
-int64_t jls_rd_fsr_utc_to_sample_id(struct jls_rd_fsr_s * self, int64_t utc) {
+int32_t jls_rd_fsr_timestamp_to_sample_id(struct jls_rd_fsr_s * self, int64_t timestamp, int64_t * sample_id) {
     if (self->entries_length == 0) {
-        return 0;
+        return JLS_ERROR_UNAVAILABLE;
     } else if (self->entries_length == 1) {
         if (self->sample_rate <= 0) {
-            return 0;
+            return JLS_ERROR_UNAVAILABLE;
         }
-        double dt = (double) (utc - self->utc[0]);
+        double dt = (double) (timestamp - self->utc[0]);
         dt *= (1.0 / JLS_TIME_SECOND);
-        return self->sample_id[0] + (int64_t) (dt * self->sample_rate);
+        *sample_id = self->sample_id[0] + (int64_t) (dt * self->sample_rate);
+    } else {
+        *sample_id = interp_i64(self, timestamp, self->utc, self->sample_id);
     }
-    return interp_i64(self, utc, self->utc, self->sample_id);
+    return 0;
 }